ch3: add command line option to plot.sh

+++ ch3/plot.sh
@@ -1,4 +1,12 @@
 #!/bin/bash
-gnuplot power_function.gp
+
+# Check if a gnuplot file argument is provided
+if [ -z "$1" ]; then
+    echo "Usage: $0 <gnuplot_script>"
+    exit 1
+fi
+
+gnuplot $1
 pdflatex --interaction=batchmode figure.tex
 rm *.aux *.log *-inc.eps *.tex *converted-to.pdf
+open figure.pdf
